backend/requirements.txt
Django>=4.2
djangorestframework
requests           # Để gọi API sang Colab
paho-mqtt          # Để giao tiếp MQTT với ESP32
django-cors-headers # Để React Native gọi API không bị chặn
python-dotenv      # Để quản lý biến môi trường
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
supabase>=0.0.1
gradio